1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include <stdio.h>
+#include <stdlib.h>
 #include "function_pointers.h"
 
 /**
@@ -10,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-unsigned int i = 0;
+size_t i = 0;
 if (array != NULL && action != NULL && size > 0)
 {
 while (i++ < size)
